Fix silent typecasting
Numpy allowed casting of single-element array to float, but doesn't like
that anymore and now warns:

DeprecationWarning: Conversion of an array with ndim > 0 to a scalar is
deprecated, and will error in future. Ensure you extract a single
element from your array before performing this operation. (Deprecated
NumPy 1.25.)

so let's get rid of those.