Make optimization more robust
Minimize functions with L-BFGS-B

We previously used COBYLA, but with the release of scipy 1.16 it became
about 30 times slower (they switched to a new implementation in pure
Python, and state that it is not well suited when function evaluation is
cheap, which is our case).

So we switch to L-BFGS-B, which seems better behaved for our use case.